Automatically add END to invocations of MACRO() and friends
Requiring end-users to terminate macros with END strikes me as
easy to mess up, and perhaps inelegant.  This commit removes the
requirement for end-users to terminate macros with END.

As a result of this commit, end-users (including other plugins) who
do use END will see a tiny amount of increased code size (1 byte per
declared macro I believe), but functionality remains intact.

Usage of END is hereby deprecated, and eventually #define END may
be removed in a future commit.

README.md has been modified with the new usage instructions, and
a note that usage of END is deprecated.