Add helpers to aid in implementing key masking
There are situations when one wants to ignore key events for a while, and mask
them out. These newly introduced functions help do that.

They are in the Hardware plugin, because this is where it is most efficient to
implement the masks: the hardware library knows how many bits it needs, and how
best to represent the masks. We use a 32-bit bitmap here, other keyboards may
use a different size, or an entirely different approach too.

This is one part of the fix to address keyboardio/Kaleidoscope#150.

Signed-off-by: Gergely Nagy <algernon@madhouse-project.org>