Held + normal key should prevent oneshot from firing
When holding a one-shot key and tapping another that would cancel the one-shot
effect, when we release the one-shot key within `hold_time_out`, do not start
the oneshot effect.

This is done by only clearing the `should_cancel_` flag in `loopHook` when
cancellation did happen. If we clear anyway, then the flag set by the
interrupting keypress will be lost by the time we release the one-shot key.

Reported by @ToyKeeper, thanks a lot for the detailed explanation!

Signed-off-by: Gergely Nagy <algernon@madhouse-project.org>