When cancelling after hold timeout, turn should_cancel_ off too.
If we do not turn `should_cancel_` off, the next time one taps a one-shot key,
the flag will still be on (because nothing else turned it off, and it was set
before `hold_time_out_` happened, because the normal one-shot timeout is shorter
than that). If the flag is on, the loop hook will turn any and all one-shot
effects off.

In practice, this resulted in one-shot keys not working properly if they were
held for a longer period before.

Fixes #12, with many thanks to @ToyKeeper for finding and reporting the issue
with reproduction steps.

Signed-off-by: Gergely Nagy <algernon@madhouse-project.org>