When checking for normal layer keys, make sure the RESERVED bit isn't set
If the `RESERVED` bit is set, that means the key is outside of the control of
the core firmware, and as such, can't be a normal layer key. As it happens,
`OSM(LeftControl)` (and other OSM keys) triggered this branch of code, and
resulted in one-shot modifiers getting highlighted too.

Thanks to @ToyKeeper for reporting the problem, and providing reproduction
steps.

Signed-off-by: Gergely Nagy <algernon@madhouse-project.org>