Update the key cache on layer change and at keyboard setup
When we change layers, we want to update the key cache for the whole keyboard,
so that LED modes and other things that depend on all keys being up-to-date will
work as expected.

Do the same at `Kaleidoscope.setup` time, so we start with a good state too.

This fixes keyboardio/Kaleidoscope-Numlock#7.

Signed-off-by: Gergely Nagy <algernon@madhouse-project.org>