Smarter Layer.on / Layer.off
Only update the keymap cache if the layer state changed for real. If we turn a
layer that was already on, on again, we do not need to update. Same for turning
them off.

This results in a tiny speedup if we have code that calls `Layer.on()` or
`Layer.off()` often.

Signed-off-by: Gergely Nagy <algernon@madhouse-project.org>