When a momentary layer is held, reactivate the layer if it turns off
If we have two keys on our keymap that momentarily go to the same layer (which
is the case for the factory firmware), we hold both, and release one, we want
the layer to remain active still.

To this effect, in `handleKeymapKeyswitchEvent` we will handle the case when a
momentary layer key is pressed, but not toggled on (that is, it is held): if it
is not a next/previous switch, we re-activate the layer if it wasn't on.

This fixes #154, thanks to @ToyKeeper for the report.

Signed-off-by: Gergely Nagy <algernon@madhouse-project.org>