Redo the logic that checks if we need to highlight
Instead of using `Kaleidoscope-Ranges` and custom helper functions with magic
constants to decide whether we need to highlight a key, refresh it, or leave it
alone, use an if-else chain and inner ifs for activity.

Leverages the new `OneShot.isOneShotKey(key)` and `OneShot.isActive(key)`
methods.

The net result is slightly cleaner code (though it can still be improved), and
about 0.2ms saved, along with some PROGMEM space.

Signed-off-by: Gergely Nagy <kaleidoscope@gergo.csillger.hu>