Major performance optimization
Now that we only care about the highest active layer, make sure we only do work
if that changed between two cycles. This way, `onActivate()` will take care of
the initial setup, `update()` will be an almost no-op for most of the time, and
`refreshAt()` will take care of refreshing keys other plugins may have changed.

This makes us about twice as fast as we were, on average.

Signed-off-by: Gergely Nagy <kaleidoscope@gergo.csillger.hu>