Layer: Add Layer.highest() to access the cached highestLayer
Sometimes we would like to know the highest active layer, which is available in
the `highestLayer` private variable, and via `.top()` too. We do not want to
make `highestLayer` public, because we want to be the only ones changing it. And
while `top()` gets us roughly the same information, it does so at a cost. For a
lot of purposes, the cached `highestLayer` would be perfectly adequate.

The new `Layer.highest()` accessor does just this.

Signed-off-by: Gergely Nagy <kaleidoscope@gergo.csillger.hu>