Add a way to optimize for a one-shot-only setup.
OneShot has a convenient way to tell us if any OneShot keys are active:
`OneShot.isActive()`. Thus, if we are using oneshots only, we can skip scanning
the whole keymap if no one-shots are active, saving us a whole lot of time per
cycle we would be spending needlessly.

The optimization is off by default, and must be turned on by the user's sketch.

Signed-off-by: Gergely Nagy <kaleidoscope@gergo.csillger.hu>