Use an event loop hook instead of requiring a macro to detect NumLock
Requiring the end-user to use a macro to have the NumLock effect is a bit
confusing. We can do better than that, by using an event handler hook, and
catching `Keypad_NumLock` presses, and toggle on keypress.

This way, the macro is not necessary, and all the user has to do, is to use the
plugin, configure `numPadLayer`, and done.

Signed-off-by: Gergely Nagy <kaleidoscope@gergo.csillger.hu>