Deprecate USE_PLUGINS
`Kaleidoscope.use` is a much better interface, therefore deprecate USE_PLUGINS.
We do this by creating a wrapper function, `__USE_PLUGINS` that will call
`Kaleidoscope.use` under the hood, but has a deprecated attribute attached. We
then make the `USE_PLUGINS` macro call this function.

We do this because we want to make sure that the list is NULL-terminated, and
for that, we need the macro.

Signed-off-by: Gergely Nagy <kaleidoscope@gergo.csillger.hu>