Add LEDControl.reactivate()
Intended to force-reactivate the current LED mode, in case we want to refresh
the whole board, and make sure we do so even if the current mode's update is a
no-op.

This can happen when we overrode some keys, and it becomes less costly to update
everything than to iterate over the updated keys.

Signed-off-by: Gergely Nagy <kaleidoscope@gergo.csillger.hu>