Introduce KaleidoscopePlugin.initialSetup, and deprecate .begin
As discussed in #196, if we are making `KaleidoscopePlugin.begin` protected, we
might as well give it a better name. That name is `initialSetup`, and this
change is the first step towards the migration. It introduces `initialSetup`
which will call `begin` for now, and deprecate `begin`, which is no longer an
abstract function.

Once everyone migrated to the new name, we can remove `.begin`, and turn
`.initialSetup` into an abstract function.

Signed-off-by: Gergely Nagy <kaleidoscope@gergo.csillger.hu>