The first half of https://github.com/keyboardio/Kaleidoscope/pull/196
Make Kaleidoscope_ a friend class, so that it can access .begin. The
reason behind this is that .begin is an interface towards
Kaleidoscope.use(), and that function should be the only user. To
discourage its use, make it protected.

This does not break any existing - and valid - code, but allows us to
slowly migrate the plugins to a protected begin() method.

Fixes #177.