ConsumerControl: Report separately from press / release
Instead of always sending a report on every `press`, `release`, and
`releaseAll`, let the user of this code decide when to send the report. This
will allow us to treat `ConsumerControl` the same way we do with `Keyboard`, and
clear the report each cycle. That in turn will help us not send many presses for
the same key in a single report.

Signed-off-by: Gergely Nagy <kaleidoscope@gergo.csillger.hu>