Fix ConsumerControl repeating
Instead of sending a press & release for the consumer key when the physical key
is released, send a press each cycle it is held, a report along with the
keyboard report, and clear the consumer report each cycle too.

This will prevent these keys getting stuck, or sending multiple presses in the
same report.

Fixes #176.

Signed-off-by: Gergely Nagy <kaleidoscope@gergo.csillger.hu>