Mouse: Pull sendReport out of move & buttons
Instead of sending a report immediately after a move or a button press, let the
user of the library decide when to send it. While there, also expose a
`releaseAll` method, so we follow the same pattern as `Keyboard`, allowing the
user to send & clear the report each cycle.

This separation is required to address keyboardio/Kaleidoscope-MouseKeys#10.

Signed-off-by: Gergely Nagy <kaleidoscope@gergo.csillger.hu>