Add `ENABLE_BOOT_KEYBOARD()`
To support the boot report protocol, we need to enable the boot keyboard. We
don't want to do this by default, because it's significant amount of code, so it
ships disabled. To enable it, we need to do a small dance, and the newly
introduced `ENABLE_BOOT_KEYBOARD` macro does just that. It also sets the default
protocol to boot.

Signed-off-by: Gergely Nagy <algernon@keyboard.io>