Mouse: Append the descriptor late, in Mouse.begin
Because `Mouse` is *always* instantiated, the constructor always runs, and thus,
the node is always registered whenever the `Mouse` object is referenced, even if
done so in code that gets optimized out. To combat this, and allow the compiler
to optimize out `Mouse` fully, do the append in `Mouse.begin`. This way, if the
`Mouse` global is referenced by code that gets optimized out, the constructor
will be optimized out, allowing the full object to get dropped. As we call
`Mouse.begin` during `setup`, there is no risk doing a late append, either.

This addresses the most important part of keyboardio/Kaleidoscope#257.

Signed-off-by: Gergely Nagy <algernon@keyboard.io>