Add a special `WakeupKeyboard`, to be able to wake the host up
At least on Linux, for a device to be considered capable of waking the host up,
it must be a boot keyboard. As we do not (yet) support a boot keyboard, we fake
one. An USB node that does nothing else than report itself as a boot keyboard,
and does the minimum amount of work to get recognised as such.

Because of this, Linux - and hopefully the other OSes too - will consider the
whole device capable of waking up the host.

This addresses keyboardio/Kaleidoscope#237, if all goes well.

Signed-off-by: Gergely Nagy <algernon@keyboard.io>