Fixed scan order bug
Store any macros key events and play them after the event handler pass has finished, so we
don't have a problem when holding other keys that are handled after the macro key in a
pass. This fixes the problem where held modifiers wouldn't be applied to macros, and also
fast repeating of printing characters.

This change does introduce a limit (default: 8) on the number of concurrent macros that
can be played.