Changed behaviour for layer-shift keys
Layer shift keys now set `highestLayer`, but don't affect `layerState`. This means that
when holding a `ShiftToLayer(N)` key, layer N becomes the top active layer, and layers
above N are ignored while that key is held. This should result in less surprising
behaviour for users who are unfamiliar with the layer stack.

In addition, this means that only one layer shift can be active at a time. It also means
that locked layers are not deactivated when a matching layer shift key is released.

To make this change, I turned `handleKeymapKeyswitchEvent()` into a private static member
function so it could access the `highestLayer()` variable. It was already hidden from
other translation units, so this shouldn't be a big deal. I also added `row` & `col` as
parameters so that the key in question could be masked if it detects a change to
`highestLayer` (e.g. from another layer shift key being pressed -- we don't want them
fighting for control).

`updateActiveLayers()` was changed so that `highestLayer` always gets a lookup, regardless
of its bit in `layerState`.