Added `Layer.shiftOn` boolean to prioritize layer shifts
With this change, when a `ShiftToLayer()` key is held, any layers that get activated or
deactivated (via `Layer.On()` & `Layer.Off()`) won't cause the layer shift to be dropped,
even if a higher layer than the shifted-to layer is activated. It will take effect once
the layer shift key is released, however.