Corrected macro for HID_CONSUMER_AC_DOWNLOAD
These macros were copied and pasted, I'm guessing from the USB HID Usage Tables document,
where this keycode is identified with a parenthetical, unlike all of the others around it:
`AC Download (Save Target As)`. When automatically converted into a preprocessor macro, it
gets a trailing underscore and an argument, which is not what we want.

This change won't actually fix anything that's currently broken, but it might matter
someday.