Further actOnMatrixScan optimizations
In `actOnMatrixScan`, there is no need to use temporary variables, we can just
pass the data directly to `actOnHalfRow`, and doing so makes the code easier to
follow.

In `actOnHalfRow`, we can further optimize things if instead of reading the Nth
bit, we always read the first, and shift the byte at the end.

All of these optimizations were done by @obra, I just wrote the commit message.

Signed-off-by: Gergely Nagy <algernon@keyboard.io>