Use the triggering coordinates when injecting events
When we activate a OneShot key, use the coordinates of the (last) physical key
that triggered it. This is done by keeping an array of positions for each
possible OneShot key. While this costs us 16 bytes of RAM and a bit of code, the
benefit is that plugins ordered after OneShot will know where OneShot was
triggered from.

Fixes #13.

Signed-off-by: Gergely Nagy <algernon@keyboard.io>