Use timers instead of ticks to time when to update the effects
Ticks depend on the speed of the main loop, and as such, are not a reliable way
to time animations. For this reason, use proper timers instead.

The update delay is set to 40ms, which appears to be a slow, relaxing animation,
and should be roughly in the ballpark the tick-based timing was, before speeding
up the main loop considerably.

Fixes #3.

Signed-off-by: Gergely Nagy <algernon@keyboard.io>