Do not deprecate kaleidoscope::Plugin::begin()
The problem with the deprecation is that even if we implement the new API, this
`begin()` method will still be inherited, and the deprecation triggered, even if
we migrated to use `init()` instead.

Because we deprecated the old class name, and the hook adding/removing
functions, that should be enough warnings to help anyone migrate to the new
APIs, and we do not need this one.

Signed-off-by: Gergely Nagy <algernon@keyboard.io>