Move OrderedPlugins to the `kaleidoscope_internal` namespace
We need `OrderedPlugins` outside of the `kaleidoscope` namespace, because
otherwise there will be conflicts between classes under `kaleidoscope`, and
instances of them in the global namespace. In practice, `OrderedPlugins` will
try to use the classes instead of the global objects, because it looks in
`kaleidoscope::` first, if it is there.

If we move it outside, it will correctly find the global objects.

Signed-off-by: Gergely Nagy <algernon@keyboard.io>