Drop the use of EventKey
Instead, use the row,col,keyState triplet like we did before. The primary reason
for this change is to ease migration, with incremental changes being a close
second. We're already redoing the API, changing it in a way that's harder to
migrate to at the same time isn't beneficial.

We may want to explore the idea at a later point, mind you.

Signed-off-by: Gergely Nagy <algernon@keyboard.io>