Disable deprecation warnings when using certain functions internally
We want to deprecate a number of functions, and emit a warning when used in user
code. On the other hand, there are legacy uses of these within Kaleidoscope
itself, which we'd rather not see warnings for. Thus, we disable the deprecation
warnings for those and those functions only, that are legitimate users of the
deprecated functions. Often, they are deprecated themselves too.

Signed-off-by: Gergely Nagy <algernon@keyboard.io>