Header movements & comments
Moved hook_rooting.h to plugin_hooks.h, and added a comment explaining why we
use the kaleidoscope_internal namespace, and why the header is in this
directory.

Also moved preprocessor_macro_map.h out of internal, and shortened its name (to
match the style we used for macro_helpers.h). I did not fold it into
macro_helpers.h, due to the different licensing. Easier to keep track of that if
it is in a separate file.

Signed-off-by: Gergely Nagy <algernon@keyboard.io>