Add const and constexpr specifiers to Key union
Making the member functions of Key `const' explicitly flags that they
will not change the union. This will allow to use Key in const contexts.

Adding the `constexpr' specifier to the function makes it possible to
rely on the results at compile time. This puts some kind of restrictions
on the function, especially when using C++11 and not a newer standard,
but these restrictions were already fulfilled, so this seems to be safe.