Turn kaleidoscope::Hooks into a namespace
We want to allow plugins to have their own hooks, without having to touch the
core firmware. For this reason, protecting the hooks via a class and friend
declarations does not make sense. Nor does a class, for that matter, as it is
not extensible from the outside. A namespace, however, is.

And this is what this change does: turns `kaleidoscope::Hooks` into a namespace
called `kaleidoscope::hooks`. This way, plugins can declare new hooks within it,
and don't need to sync with core.

Fixes #323.

Signed-off-by: Gergely Nagy <algernon@keyboard.io>