Dim the numpad color down from 255 to 160.
Having it at the brightest uses too much power, and may result in power use
surges when switching between LED modes and NumPad, which in turn can force the
operating system to disable the whole device. To avoid this, lower the
brightness to 160, a carefull tuned value, also used by the `solidRed` mode in
the factory firmware.

Fixes #9.

Signed-off-by: Gergely Nagy <algernon@keyboard.io>