Make the V1 plugin API ifdefs broader
Instead of only having an ifdef in the body of deprecated methods, and leaving
it up to the compiler to optimize out the empty & unused, explicitly wrap the
declaration of them within an ifdef too. This will make it easier to remove
everything V1 at a later point, and we're not at the mercy of the compiler,
either.

Fixes #327.

Signed-off-by: Gergely Nagy <algernon@keyboard.io>