Clear should_cancel_ when canceling stickiness with a third tap
If we do not clear `should_cancel_` there, it remains set indefinitely (until it
gets unset for some other reason, like interrupting a non-sticky OneShot). If it
is set, we can't enter sticky mode.

So clear it when canceling stickiness with a third tap, just like we clear it in
`afterEachCycle`.

Fixes #34.

Signed-off-by: Gergely Nagy <algernon@keyboard.io>