Fix the way we trigger an error on Kaleidoscope.use() with V2 API
When using `Kaleidoscope.use()` and the V1 API is disabled, we want to display
an error. The current method of doing that is not reliable, it sometimes works,
sometimes will error out even when not using `Kaleidoscope.use()`. To fix this,
delay the initialisation of `.use()`, so it only evaluates when used, and thus,
only fails with a descriptive error in that case.

Signed-off-by: Gergely Nagy <algernon@keyboard.io>