If a non-Topsy key toggles on, disable repeating the Topsy one
This is to avoid the case where chording a Topsy and a non-Topsy key would
result in both being repeated very fast. This happened because each cycle we
sent two reports: one for the Topsy key, one for the others. This didn't allow
the host's normal repeating behaviour to kick in, and instead, resulted in both
keys being input very fast.

Fixes #5.

Signed-off-by: Gergely Nagy <algernon@keyboard.io>