Seal the EEPROM layout automatically
If `EEPROMSettings.seal()` wasn't explicitly called, seal the layout in
`beforeEachCycle()`. On the flip side, this makes user sketches simpler, because
they don't have to seal explicitly. This is done at the cost of an if check each
cycle.

In the long run, EEPROM layout management will be moving out of this plugin, so
this check will be eventually dropped too.

Signed-off-by: Gergely Nagy <algernon@keyboard.io>