Try to avoid a possible PROGMEM overflow in getKeyOverride
If we're looking up a key from `PROGMEM`, only do that if the layer in question
is smaller than `layer_count`. Doing otherwise would read garbage from `PROGMEM`
in case we try to read from a layer higher than what we have in there. This can
happen if we have more layers in `EEPROM` than in `PROGMEM`.

Signed-off-by: Gergely Nagy <algernon@keyboard.io>