Add support for saving the default layer
With upcoming EEPROMKeymap changes, saving the default layer will become an
important part of the keyboard settings. The new `default_layer` method provides
a way to set it programmatically, while the `settings.defaultLayer` Focus
command allows the user to set it via Focus.

The default layer - if set - is activated when the settings are sealed, either
directly or automatically.

Signed-off-by: Gergely Nagy <algernon@keyboard.io>